i32 -> i64

+++ src/interpreter/bytecode/compiler/compiler.rs
@@ -532,7 +532,7 @@ mod tests {
   fn binary_operators() {
     let decl = parse_fn_decl(
       r#"
-      fn f() -> i32 {
+      fn f() -> i64 {
         ret 2 * (3 + 4)
       }
       "#
@@ -585,7 +585,7 @@ mod tests {
   fn if_statement() {
     let decl = parse_fn_decl(
       r#"
-      fn f() -> i32 {
+      fn f() -> i64 {
         if 0 {
           ret 1
         } else {
@@ -638,7 +638,7 @@ mod tests {
   fn loop_statement() {
     let decl = parse_fn_decl(
       r#"
-      fn f() -> i32 {
+      fn f() -> i64 {
         loop {
           if x {
             break
@@ -698,7 +698,7 @@ mod tests {
   fn lexical_scoping() {
     let decl = parse_fn_decl(
       r#"
-      fn f() -> i32 {
+      fn f() -> i64 {
         let x = 1
         {
           let x = 2
@@ -759,7 +759,7 @@ mod tests {
   fn store_and_load_local() {
     let decl = parse_fn_decl(
       r#"
-      fn f() -> i32 {
+      fn f() -> i64 {
         let x = 1 + 2
         ret x
       }
@@ -787,7 +787,7 @@ mod tests {
   fn store_and_load_mutable_local() {
     let decl = parse_fn_decl(
       r#"
-      fn f() -> i32 {
+      fn f() -> i64 {
         mut x = 1 + 2
         ret x
       }
@@ -815,7 +815,7 @@ mod tests {
   fn rebind_mutable_local() {
     let decl = parse_fn_decl(
       r#"
-      fn f() -> i32 {
+      fn f() -> i64 {
         mut x = 1
         x = 70
         ret x
@@ -844,7 +844,7 @@ mod tests {
   fn cannot_rebind_immutable_local() {
     let decl = parse_fn_decl(
       r#"
-      fn f() -> i32 {
+      fn f() -> i64 {
         let x = 1
         x = 70
         ret x
@@ -861,7 +861,7 @@ mod tests {
   fn cannot_rebind_unknown_local() {
     let decl = parse_fn_decl(
       r#"
-      fn f() -> i32 {
+      fn f() -> i64 {
         mut x = 1
         y = 70
         ret x
@@ -937,7 +937,7 @@ mod tests {
   fn fn_decl_with_parameters() {
     let decl = parse_fn_decl(
       r#"
-      fn f(a: i32, b: i32, c: i32) {
+      fn f(a: i64, b: i64, c: i64) {
         ret a + (b + c)
       }
       "#
@@ -998,7 +998,7 @@ mod tests {
   fn simple_tail_recursion_no_args() {
     let decl = parse_fn_decl(
       r#"
-      fn f() -> i32 {
+      fn f() -> i64 {
         ret f()
       }
       "#
@@ -1019,7 +1019,7 @@ mod tests {
   fn simple_tail_recursion_with_args() {
     let decl = parse_fn_decl(
       r#"
-      fn f(a: i32, b: i32) -> i32 {
+      fn f(a: i64, b: i64) -> i64 {
         ret f(a, b)
       }
       "#
